Test Ahoy.geocode
Test that setting Ahoy.geocode to false removes the three
location-specific keys from the VisitProperties hash.

Prior to this commit there were no tests, so this commit introduces a
simple test harness using Minitest (as per
https://github.com/ankane/searchkick/). Since this gem is designed to be
used inside a Rails application it makes use of methods that only rails
provides (notably for this commit VisitProperties uses `delegate` from
active_support). This meant I had to introduce rails as a development
dependency in order for the tests to run.